changes in rawdata_tap
